refactor: update folder location #31
